서블릿 11일차 : 게시글 서비스 예외처리

+++ servlet/servlet_project2/src/main/java/kr/kh/app/service/BoardServiceImp.java
@@ -169,15 +169,18 @@ public class BoardServiceImp implements BoardService {
 		   return false;
 		}
 		
-		//삭제할 첨부파일 삭제
-		for(String numStr : nums) {
-			try {
-				int num = Integer.parseInt(numStr); //정수로 변환
-				FileVO fileVo = boardDao.selectFile(num); //기본키로 전달할 것이기에 list 노노
-				deleteFile(fileVo);
-			}catch(Exception e) {
-				e.printStackTrace();
-				
+		if(nums != null) {
+		
+			//삭제할 첨부파일 삭제
+			for(String numStr : nums) {
+				try {
+					int num = Integer.parseInt(numStr); //정수로 변환
+					FileVO fileVo = boardDao.selectFile(num); //기본키로 전달할 것이기에 list 노노
+					deleteFile(fileVo);
+				}catch(Exception e) {
+					e.printStackTrace();
+					
+				}
 			}
 		}
